feat(candy): correctly implement the candy command in nu.
The candy command is the command that I have aliased/defined as my dotfiles
management interface, as described in the following article: [Atlassian
Dotfiles](https://www.atlassian.com/git/tutorials/dotfiles).

This implements it in the standard git-wrapping interface, although this is
subject to change in the near future with the Rust release of dotcandyd (date of
release yet unknown).