fix(auth): Update the keychain procedure and associated bw call
This allows a more seamless use of the bw command in the initialization of
direnv, and the ssh keychain, along with anything else. Because we define our
auth token in our secret envrc, we need to make sure that our setting the
important variable happens before our call to bw, which is only guaranteed if we
make it to the prompt at first, due to the way nushell hooks direnv.