feat(auth): Alias the unlock of bw via env-variable for consistent use
There are a few programs that I have been working with that could use this
update. This keeps the manager unlocked whenever the shell is opened by storing
the password in the envrc.secrets file. This should probably change, if I can
use an API key that would be ideal.