feat(scripts): create a submodule for the community-created scripts repository
The repository is https://github.com/nushell/nu_scripts, and it contains a lot
of potentially useful things to include in any given nushell configuration. This
commit brings in the outside repository as a submodule for my own personal
configuration repository. With just this commit, the scripts won't work though,
we need an additional step...