feat(pgp): include the necessary export of GPG_TTY during nu config
I wanted to put this in with the direnv files or in login.nu so that it
wouldn't trigger each time and had some amount of credential memory.
that doesn't work it seems, so back to the config file it goes. this is
necessary to allow for external programs to be able to correctly link to
the gpg-agent without error.