fix(raur): provide start of modal search subcommand
This subcommand will be something somewhat like the `raur install` definition,
but with the added step of providing a menu for the user to select their desired
mode to use with the selected packages. In this sense, `raur install` will
become something like a shortcut to using `raur search` in install mode.