wip(mnt): update `mnt` drive credential parameter handling
This is not quite finished, but the idea here is that we want to be able to
mount and unmount crucial external disk resources in an ergonomic way. In order
to support this effort, the mnt command needs to have a well-defined method of
parsing and handling the different credential options available (password,
keyfile, or FIDO2 credential, potentially GPG key if very necessary).