feat(std): introduce `coalesce` and `spread` modules
The `coalesce` module defines a command which can be used like the coalesce
function in the Microsoft Power Suite of Products (e.g. PowerBI). Given
arguments of arbitrary size, shape, or cardinality, `coalesce` returns the first
non-null value.

The `spread` module is still WIP but the goal here is to allow spreading of a
record item in nushell into key-value flag assignments for a command.